build.py: put copy_file_if_different in separate function and use for Win and OSX
On OSX everything was fine codewise, but the same logic wasn't applied on Windows.
On Windows there was still a bug if you updated sokol and didn't delete the DLLs in the
executables working directory. You then executed with an old DLL and got weird errors.
This only happened for the hot-reload, not release builds.